3: Extract agnostic spec
